Split types and lib files
