+++ src/pages/projects/index.tsx
@@ -12,8 +12,8 @@ import { SEO } from '~/components/seo';
 import { Spacer } from '~/components/Spacer';
 import { TechStack } from '~/components/TechStack';
 
-import { getAllProjects } from '~/lib/content';
-import type { Project } from '~/types/content';
+import { getAllProjects } from '~/lib/project';
+import type { Project } from '~/types/project';
 
 export default function ProjectsIndexPage({
   projects,
